Arrange the elements using flex-box
I made each section of the html, incluiding the body, to be a flex container.
I discovered that extensions can mess up things, taking space from the end of the body,
so I set the .container (central section) top and bottom margins to auto. That pulls the
footer to the end, keeping the .container in the middle of the page.
The body and .container have a column direction, and the .header and the .footer (and their
lists) have a row direction.