Put the footer to bottom and made the card sizes more flexible
I set the justify-content of the body to space-between, so the footer is
at the flex-end, but the .container was at the middle of the page with
space between it and the header/footer. The solution to this was allowing
the .container to grow, so the extra height was absorbed by its flex-items.

The problem was that each card also absorbed that height, getting a lot
of ugly void space. It was solved setting the .cards_container height to
fit-content. That way, only the .sidebar absorbed the extra height.

I also changed the cards' size because, when the page's width was just
1px smaller than the .cards_container's, last card of each row wrapped,
letting a lot of empty space in the way.
Now, each card's size is set to 250px (through flex-basis) and grows to
the 300px max-width. That way, when the page is resized, each card can
loose 50px of width, after that, the last of the row wraps and everyone
can grow again to 300px. This is a lot more flexible and aesthetic.